perwez/pyproject.toml
[tool.poetry]
name = "perwez"
version = "0.1.1"
description = ""
authors = ["reth <sosp20212021@gmail.com>"]

[tool.poetry.dependencies]
python = ">=3.6 <3.10"
netifaces = "^0.10.9"
psutil = "^5.7.2"
tenacity = "^6.2.0"
aiohttp = "^3.7.2"
portpicker = "^1.3.1"
loguru = "^0.5.3"
msgpack = "^1.0.2"
lz4 = "^3.1.1"
pyzmq = "^20.0.0"
requests = "^2.25.1"

[tool.poetry.dev-dependencies]
black = "^20.8b1"
pylint = "^2.6.0"
pytest = "^6.1.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
